Extract layout component
